Add websocket client transport
`service.Client` now accepts URIs of the scheme `ws://...` and
`wss://...` on the `Connect` function.  By default this will set the
Origin header to the Host component of the URI, however this may be
overridden by setting the `WebsocketOrigin` field of `service.Client` to
the desired value.  Further control over the `websocket.Dialer` used to
establish the connection may be had by obtaining a Dialer via the
`NewWebsocketDialer()` function (which will set sensible defaults for
use with surgemq), and assigning it to the `WebsocketDialer` field of
the Client.